@@ -5,12 +5,11 @@ homepage: https://z.flutter.gallery/
 repository: https://github.com/jamesblasco/zflutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  quiver: ^2.1.3
 
 dev_dependencies:
   flutter_test:
