@@ -24,8 +24,10 @@ class _ModalBottomSheetExampleState extends State<ModalBottomSheetExample> {
         expand: true,
         context: context,
         backgroundColor: Colors.transparent,
-        builder: (context, scrollController) => Close(
-          child: PhotoShareBottomSheet(scrollController: scrollController),
+        builder: (context) => Close(
+          child: PhotoShareBottomSheet(
+            scrollController: ModalScrollController.of(context)!,
+          ),
         ),
       );
     });
@@ -78,8 +80,9 @@ class _ModalBottomSheetExampleState extends State<ModalBottomSheetExample> {
                 expand: true,
                 context: context,
                 backgroundColor: Colors.transparent,
-                builder: (context, scrollController) =>
-                    PhotoShareBottomSheet(scrollController: scrollController),
+                builder: (context) => PhotoShareBottomSheet(
+                  scrollController: ModalScrollController.of(context)!,
+                ),
               );
             },
           ),
@@ -100,7 +103,7 @@ class _ModalBottomSheetExampleState extends State<ModalBottomSheetExample> {
 class Close extends StatefulWidget {
   final Widget child;
 
-  const Close({Key key, this.child}) : super(key: key);
+  const Close({Key? key, required this.child}) : super(key: key);
   @override
   State<StatefulWidget> createState() => CloseState();
 }
