@@ -1,3 +1,4 @@
+//@dart=2.12
 import 'package:flutter/rendering.dart';
 
 import '../core.dart';
@@ -10,6 +11,11 @@ class RenderZBox extends RenderBox {
   void performSort() {
     sortValue = this.origin.z;
   }
+
+  @override
+  Size computeDryLayout(BoxConstraints constraints) {
+    return constraints.biggest;
+  }
 }
 
 enum SortMode { inherit, stack, update }
@@ -19,8 +25,8 @@ class RenderZMultiChildBox extends RenderZBox
         ContainerRenderObjectMixin<RenderZBox, ZParentData>,
         RenderBoxContainerDefaultsMixin<RenderZBox, ZParentData> {
   RenderZMultiChildBox({
-    List<RenderZBox> children,
-    SortMode sortMode = SortMode.inherit,
+    List<RenderZBox>? children,
+    SortMode? sortMode = SortMode.inherit,
   })  : assert(sortMode != null),
         this.sortMode = sortMode {
     addAll(children);
@@ -42,7 +48,8 @@ class RenderZMultiChildBox extends RenderZBox
     super.performLayout();
 
     final BoxConstraints constraints = this.constraints;
-    RenderZBox child = firstChild;
+
+    RenderZBox? child = firstChild;
 
     while (child != null) {
       final ZParentData childParentData = child.parentData as ZParentData;
@@ -51,7 +58,12 @@ class RenderZMultiChildBox extends RenderZBox
       } else {
         child.layout(constraints, parentUsesSize: false);
       }
-      child = childParentData?.nextSibling;
+
+      /*  final Size childSize = child.size;
+        width = math.max(width, childSize.width);
+        height = math.max(height, childSize.height);*/
+
+      child = childParentData.nextSibling;
     }
     performSort();
   }
@@ -66,7 +78,7 @@ class RenderZMultiChildBox extends RenderZBox
   void performSort() {
     if (sortMode == SortMode.stack || sortMode == SortMode.update) {
       final children = getFlatChildren();
-      sortValue = children.fold(0,
+      sortValue = children.fold<double>(0,
               (previousValue, element) => previousValue + element.sortValue) /
           children.length;
     } else {
@@ -78,12 +90,12 @@ class RenderZMultiChildBox extends RenderZBox
   @override
   bool get sizedByParent => true;
 
-  SortMode sortMode;
+  SortMode? sortMode;
 
   List<RenderZBox> getFlatChildren() {
     List<RenderZBox> children = [];
 
-    RenderZBox child = firstChild;
+    RenderZBox? child = firstChild;
 
     while (child != null) {
       final ZParentData childParentData = child.parentData as ZParentData;
@@ -98,7 +110,7 @@ class RenderZMultiChildBox extends RenderZBox
         width = math.max(width, childSize.width);
         height = math.max(height, childSize.height);*/
 
-      child = childParentData?.nextSibling;
+      child = childParentData.nextSibling;
     }
     return children;
   }
@@ -122,7 +134,7 @@ class RenderZMultiChildBox extends RenderZBox
     }
   }
 
-  bool defaultHitTestChildren(BoxHitTestResult result, { Offset position }) {
+  bool defaultHitTestChildren(BoxHitTestResult result, { required Offset position }) {
     if (sortMode == SortMode.inherit) return false;
     // The x, y parameters have the top left of the node's box as the origin.
     List<RenderZBox> children = getFlatChildren();
@@ -151,7 +163,7 @@ class RenderZMultiChildBox extends RenderZBox
   }
   
   @override
-  bool hitTest(BoxHitTestResult result, {Offset position}) {
+  bool hitTest(BoxHitTestResult result, {required Offset position}) {
     if (hitTestChildren(result, position: position) || hitTestSelf(position)) {
       result.add(BoxHitTestEntry(this, position));
       return true;
@@ -160,7 +172,7 @@ class RenderZMultiChildBox extends RenderZBox
   }
 
   @override
-  bool hitTestChildren(BoxHitTestResult result, { Offset position }) {
+  bool hitTestChildren(BoxHitTestResult result, { required Offset position }) {
     return defaultHitTestChildren(result, position: position);
   }
 }
@@ -177,7 +189,7 @@ class ZParentData extends ContainerBoxParentData<RenderZBox> {
     this.rotate = ZVector.zero,
     this.scale = ZVector.identity,
     this.translate = ZVector.zero,
-    List<ZTransform> transforms,
+    List<ZTransform>? transforms,
   }) : this.transforms = transforms ?? [];
 
   ZParentData clone() => ZParentData(
